Fix showing token prompt if only client_id is missing
This prevents showing of token prompt when a user has lying around a non valid
beatport_token.json in the config dir but has valid username and password
configured in the beets config. The only thing that's missing in such a
situation is a valid client_id which will now be automatically fetched.